feat: clean up metrics
