+++ dashboard/app.py
@@ -3,82 +3,27 @@ Streamlit dashboard for apple-health-data
 """
 
 import logging
-from dataclasses import dataclass
 from datetime import datetime, timedelta
-from pathlib import Path
-from typing import Any, Optional
 
 import conf
 import polars as pl
 import streamlit as st
-import yaml
-from graphing import MACROS_BAR_HEIGHT, render_macros_bar_chart
+from graphing import (
+    MACROS_BAR_HEIGHT,
+    filter_metrics,
+    render_macros_bar_chart,
+)
 from helpers import (
     compute_avg_sleep_time_from_midnight,
     convert_column_to_timezone,
     get_average,
     read_parquet_from_s3,
 )
+from kpi import load_kpi_config, render_kpis
 
 # Configure logging.
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-
-# ---------------------- KPI CONFIG ----------------------
-@dataclass
-class KPI:
-    label: str
-    key: str
-    format: str = "{}"
-    goal: Optional[float] = None
-    is_time: bool = False
-
-
-def load_kpi_config(
-    yaml_path: Path = Path(conf.kpi_config_path),
-) -> dict[str, list[KPI]]:
-    with Path.open(yaml_path) as f:
-        raw = yaml.safe_load(f)
-    return {section: [KPI(**item) for item in items] for section, items in raw.items()}
-
-
-def render_kpis(section: str, values: dict[str, Any], config: dict[str, Any]) -> None:
-    """
-    Render a section of KPI metrics to the Streamlit UI.
-
-    Args:
-        section (str): The section name (e.g. "macros", "sleep") to pull config for.
-        values (dict[str, Any]): Dictionary of computed KPI values keyed by metric name.
-        config (dict[str, Any]): Loaded YAML config mapping sections to KPI definitions.
-
-    This function looks up the metrics for the given section, formats each one according
-    to the provided rules (e.g., formatting, delta, time display), and renders them as
-    Streamlit metrics.
-    """
-    kpis = config.get(section, [])
-    cols = st.columns(len(kpis))
-
-    for col, kpi in zip(cols, kpis):
-        key = kpi.key
-        val = values.get(key)
-        fmt = kpi.format
-        is_time = kpi.is_time
-
-        goal = kpi.goal
-        delta = None
-        if goal is not None and val is not None and not is_time:
-            pct_change = (val - goal) / goal * 100
-            delta = f"{pct_change:+.0f}%"
-
-        if is_time and isinstance(val, datetime):
-            val_str = val.strftime(fmt)
-        elif val is not None:
-            val_str = fmt.format(val)
-        else:
-            val_str = "N/A"
-
-        col.metric(kpi.label, val_str, delta)
 
 
 # Page configuration
@@ -173,22 +118,13 @@ except Exception as e:
 
 # Load required macros data
 try:
-    macros_columns = [
-        "carbohydrates",
-        "protein",
-        "total_fat",
-    ]
-    _bar_macros = filtered_macros.filter(pl.col("metric_name").is_in(macros_columns))
-
-    bar_macros = _bar_macros.with_columns(
-        pl.col("metric_name").replace("total_fat", "fat").alias("metric_name")
-    ).sort(pl.col(["metric_date", "metric_name"]))
-
-    calories_df = filtered_macros.filter(pl.col("metric_name") == "calories").sort(
-        "metric_date"
+    macros_df = filter_metrics(
+        df=filtered_macros,
+        metrics=["carbohydrates", "protein", "total_fat"],
+        rename_map={"total_fat": "fat"},
     )
-    if calories_df.is_empty():
-        st.write("No calories data available for the selected date range.")
+    calories_df = filter_metrics(filtered_macros, metrics=["calories"])
+
 except Exception as e:
     st.error(f"Error generating data for macros chart: {e}")
 
@@ -197,7 +133,7 @@ st.header("Macros Breakdown Bar Chart")
 col1, col2 = st.columns(2)
 with col1:
     st.bar_chart(
-        bar_macros,
+        macros_df,
         x="metric_date",
         y="quantity",
         color="metric_name",
@@ -210,7 +146,7 @@ with col1:
 with col2:
     # === LINE GRAPH: DAILY CALORIES using st.line_chart ===
     # st.header("Daily Calories Line Chart")
-    render_macros_bar_chart(bar_macros)
+    render_macros_bar_chart(macros_df)
 
 st.header("Calories")
 st.line_chart(
@@ -230,8 +166,8 @@ try:
 
     # Convert sleep_times to Melbourne timezone
     sleep_times_local = convert_column_to_timezone(filtered_sleep_times, "sleep_times")
-    sleep_start_df = sleep_times_local.filter(pl.col("metric_name") == "sleep_start")
-    sleep_end_df = sleep_times_local.filter(pl.col("metric_name") == "sleep_end")
+    sleep_start_df = filter_metrics(filtered_sleep_times, ["sleep_start"])
+    sleep_end_df = filter_metrics(filtered_sleep_times, ["sleep_end"])
 
     # Compute averages
     avg_sleep_start = compute_avg_sleep_time_from_midnight(sleep_start_df)
